+++ init.php
@@ -1,8 +1,8 @@
 <?php defined('SYSPATH') or die('No direct script access.');
 
-if (trim(Request::detect_uri(), '/') == 'restify/test')
+if (in_array(trim(Request::detect_uri(), '/'), array('restify/test', 'restify/test/raw')))
 {
-	Route::set('restify/test', '<directory>(/<controller>)')
+	Route::set('restify/test', '<directory>(/<controller>(/<raw>))')
 		->defaults(array
 		(
 			'directory' 	=> 'restify',